move logic to start the TractActivity to base-tool module
