+++ library/base/src/main/java/org/cru/godtools/base/Constants.kt
@@ -5,5 +5,6 @@ import android.net.Uri
 // common extras
 const val EXTRA_TOOL = "tool"
 const val EXTRA_LANGUAGE = "language"
+const val EXTRA_LANGUAGES = "languages"
 
 val URI_SHARE_BASE: Uri = Uri.parse("https://knowgod.com/")
