+++ ui/base-tool/src/main/java/org/cru/godtools/base/tool/Activities.kt
@@ -1,0 +1,28 @@
+package org.cru.godtools.base.tool
+
+import android.app.Activity
+import android.content.Context
+import android.content.Intent
+import android.os.Bundle
+import java.util.Locale
+import org.ccci.gto.android.common.util.os.putLocaleArray
+import org.cru.godtools.base.EXTRA_LANGUAGES
+import org.cru.godtools.base.EXTRA_TOOL
+
+// region TractActivity
+const val EXTRA_SHOW_TIPS = "org.cru.godtools.tract.activity.TractActivity.SHOW_TIPS"
+
+fun Activity.startTractActivity(toolCode: String, vararg languages: Locale?, showTips: Boolean) =
+    startActivity(createTractActivityIntent(toolCode, *languages, showTips = showTips))
+
+fun Context.createTractActivityIntent(toolCode: String, vararg languages: Locale?, showTips: Boolean = false) =
+    Intent().setClassName(this, "org.cru.godtools.tract.activity.TractActivity")
+        .putExtra(EXTRA_TOOL, toolCode)
+        .putLanguagesExtra(*languages)
+        .putExtra(EXTRA_SHOW_TIPS, showTips)
+
+private fun Intent.putLanguagesExtra(vararg languages: Locale?) = putExtras(Bundle().apply {
+    // XXX: we use singleString mode to support using this intent for legacy shortcuts
+    putLocaleArray(EXTRA_LANGUAGES, languages.filterNotNull().toTypedArray(), true)
+})
+// endregion TractActivity
