+++ ui/shortcuts/build.gradle
@@ -17,7 +17,7 @@ dependencies {
 
     implementation project(':ui:article-renderer')
     implementation project(':ui:base')
-    implementation project(':ui:tract-renderer')
+    implementation project(':ui:base-tool')
 
     implementation "androidx.core:core-ktx:${deps.androidX.core}"
